+++ tests/release_candidate_scraper/test_release_candidate_scraper.py
@@ -12,6 +12,7 @@ from mozdownload import ReleaseCandidateScraper
 from mozdownload.utils import urljoin
 import mozhttpd_base_test as mhttpd
 
+
 firefox_tests = [
     # -p win32 -v 21.0
     {'args': {'platform': 'win32',
@@ -68,7 +69,7 @@ firefox_tests = [
               'is_stub_installer': True,
               'platform': 'win32',
               'version': '21.0'},
-     'target': 'firefox-21.0-build3.en-US.win32.exe',
+     'target': 'firefox-21.0-build3.en-US.win32-stub.exe',
      'target_url': 'firefox/candidates/21.0-candidates/build3/win32/en-US/Firefox Setup Stub 21.0.exe'},
     # -p win64 -v 37.0b1
     {'args': {'platform': 'win64',